% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_sgt.R, R/constrained_skewt.R,
%   R/sgtfamily.R, R/skew_ged.R, R/skewt-family.R, R/sym_gt.R
\name{log_lik_constrained_sgt}
\alias{log_lik_constrained_sgt}
\alias{posterior_predict_constrained_sgt}
\alias{posterior_epred_constrained_sgt}
\alias{log_lik_constrained_skew_t}
\alias{posterior_predict_constrained_skew_t}
\alias{posterior_epred_constrained_skew_t}
\alias{log_lik_sgt}
\alias{posterior_predict_sgt}
\alias{posterior_epred_sgt}
\alias{log_lik_skew_ged}
\alias{posterior_predict_skew_ged}
\alias{posterior_epred_skew_ged}
\alias{log_lik_skew_t}
\alias{posterior_predict_skew_t}
\alias{posterior_epred_skew_t}
\alias{log_lik_sym_gt}
\alias{posterior_predict_sym_gt}
\alias{posterior_epred_sym_gt}
\title{Downstream methods for sgt family}
\usage{
log_lik_constrained_sgt(i, prep)

posterior_predict_constrained_sgt(i, prep, ...)

posterior_epred_constrained_sgt(prep)

log_lik_constrained_skew_t(i, prep)

posterior_predict_constrained_skew_t(i, prep, ...)

posterior_epred_constrained_skew_t(prep)

log_lik_sgt(i, prep)

posterior_predict_sgt(i, prep, ...)

posterior_epred_sgt(prep)

log_lik_skew_ged(i, prep)

posterior_predict_skew_ged(i, prep, ...)

posterior_epred_skew_ged(prep)

log_lik_skew_t(i, prep)

posterior_predict_skew_t(i, prep, ...)

posterior_epred_skew_t(prep)

log_lik_sym_gt(i, prep)

posterior_predict_sym_gt(i, prep, ...)

posterior_epred_sym_gt(prep)
}
\arguments{
\item{i}{iteration}

\item{prep}{a brms prepared object}
}
\description{
Some prediction methods for brms sgt family. For internal use of brms methods only
}
\seealso{
\link[brms]{log_lik} \link[brms]{posterior_predict} \link[brms]{posterior_epred}
}
