% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfuns.R
\name{expose_sgt_stanvar}
\alias{expose_sgt_stanvar}
\title{Expose stanvars of SGT and Skew T for brm call}
\usage{
expose_sgt_stanvar(raw = FALSE)
}
\arguments{
\item{raw}{(boolean, default=FALSE) whether to expose raw Stan string, or brms stanvar object. The former can be used for Rstan object.}
}
\value{
stanvar object, or if raw=TRUE, a string
}
\description{
This function exposes brms stanvar object for brm call. Adding stanvar to the conventional call in brms is analogous to an _sgt call.
}
