% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgtfamily.R, R/skewt-family.R
\name{log_lik_sgt}
\alias{log_lik_sgt}
\alias{posterior_predict_sgt}
\alias{posterior_epred_sgt}
\alias{log_lik_skew_t}
\alias{posterior_predict_skew_t}
\alias{posterior_epred_skew_t}
\title{Downstream methods for sgt family}
\usage{
log_lik_sgt(i, prep)

posterior_predict_sgt(i, prep, ...)

posterior_epred_sgt(prep)

log_lik_skew_t(i, prep)

posterior_predict_skew_t(i, prep, ...)

posterior_epred_skew_t(prep)
}
\arguments{
\item{i}{iteration}

\item{prep}{a brms prepared object}
}
\description{
Some prediction methods for brms sgt family. For internal use of brms methods only
}
\seealso{
[brms:loglik()], [brms:posterior_predict()], [brms:posterior_epred()]
}
