% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgtfamily.R, R/skewt-family.R
\name{sgt}
\alias{sgt}
\alias{skew_t}
\title{Skew generalised T family for brms model}
\usage{
sgt(
  link_mu = "identity",
  link_sigma = "identity",
  link_p = "log",
  link_q = "log",
  link_lambdap1half = "logit"
)

skew_t(
  link_mu = "identity",
  link_sigma = "identity",
  link_q = "log",
  link_lambdap1half = "logit"
)
}
\arguments{
\item{link_mu}{(character, default: identity) link function for the mode of SGT distribution}

\item{link_sigma}{(character, default: exp)  link function for the scale of SGT distribution}

\item{link_p, link_q}{(character, default: exp)  link function for the kurtosis parameters of SGT distribution}

\item{link_lambdap1half}{(character, default: logit) link function for the skew parameter of SGT distribution.
Note that the actual skew parameter of SGT is lambda where -1 < lambda < 1.
A proper link function shall be \emph{tanh}.
However, due to brms's limit options for link function, we lambda to lambdap1half = \eqn{\frac{lambda+1}{2}} and use the logit link.
Hence the name "lambda p(lus) 1 then half".}
}
\value{
a `brms::custom_family` object
}
\description{
Create a custom family object for brms model
}
\examples{
x <- rnorm(1000, 0, 1)
y <- 3*x + 3*rt(1000,3,0)
model <- brm_sgt(y ~ x, family=sgt(), prior=c(sgt_default_prior()), data=data.frame(x=x, y=y), chains=1, iter=1000)
}
\seealso{
\link[brms]{dsgt}
}
