% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_sgt.R, R/constrained_skewt.R,
%   R/sgtfamily.R, R/skewt-family.R
\name{constrained_sgt}
\alias{constrained_sgt}
\alias{constrained_skew_t}
\alias{sgt}
\alias{skew_t}
\title{Skew generalised T family for brms model}
\usage{
constrained_sgt(
  link = "identity",
  link_sigma = "log",
  link_qinv = "log",
  link_pq = "logm1",
  link_lambdap1half = "logit"
)

constrained_skew_t(
  link = "identity",
  link_sigma = "log",
  link_qmhalf = "log",
  link_lambdap1half = "logit"
)

sgt(
  link = "identity",
  link_sigma = "log",
  link_p = "log",
  link_q = "log",
  link_lambdap1half = "logit"
)

skew_t(
  link = "identity",
  link_sigma = "log",
  link_q = "log",
  link_lambdap1half = "logit"
)
}
\arguments{
\item{link}{(character, default: identity) link function for the mode of SGT distribution}

\item{link_sigma}{(character, default: log)  link function for the scale of SGT distribution}

\item{link_pq}{(string, default: logm1) link function for pq = p*q, must be > 1.}

\item{link_lambdap1half}{(character, default: logit) link function for the skew parameter of SGT distribution.
Note that the actual skew parameter of SGT is lambda where -1 < lambda < 1.
A proper link function shall be \emph{tanh}.
However, due to brms's limit options for link function, we lambda to lambdap1half = \eqn{\frac{lambda+1}{2}} and use the logit link.
Hence the name "lambda p(lus) 1 then half".}

\item{link_qmhalf}{(string, default: log) link function for qmhalf = q-1/2. Must be > 0.}

\item{link_p, link_q}{(character, default: log)  link function for the kurtosis parameters of SGT distribution}
}
\value{
a `brms::custom_family` object
}
\description{
Create a custom family object for brms model
}
\details{
The constrained_ versions constraint p*q > 1 to ensure identifiability of the mean,
but not as "generalised" due to the exclusion of the Cauchy-related branch.
The mu of the param is still the mode though.
}
\examples{
x <- rnorm(1000, 0, 1)
y <- 3*x + 3*rt(1000,3,0)
model <- brm_sgt(y ~ x, family=sgt(), prior=c(sgt_default_prior()), data=data.frame(x=x, y=y), chains=1, iter=1000)
}
\seealso{
\link[brms]{dsgt}
}
