% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_sgt.R, R/constrained_skewt.R,
%   R/sgtfamily.R, R/skewt-family.R, R/sym_gt.R
\name{constrained_sgt_default_prior}
\alias{constrained_sgt_default_prior}
\alias{constrained_skew_t_default_prior}
\alias{sgt_default_prior}
\alias{skew_t_default_prior}
\alias{sym_gt_default_prior}
\title{Default priors for sgt family}
\usage{
constrained_sgt_default_prior(params = "all", exclude = FALSE)

constrained_skew_t_default_prior(params = "all", exclude = FALSE)

sgt_default_prior(params = "all", exclude = FALSE)

skew_t_default_prior(params = "all", exclude = FALSE)

sym_gt_default_prior(params = "all", exclude = FALSE)
}
\arguments{
\item{params}{string vector (default="all") of params in sgt family, namely `sigma`, `lambdap1half`, `p`, `q`, a special value "all" will return priors for all.}

\item{exclude}{(boolean, default=FALSE) whether to include or exclude parameters in `params`.}
}
\value{
object of class `brmsprior`
}
\description{
This function returns default set of prior for sgt fit. By default, returns prior for all parameters.
If you set some parameter sub-model, please exclude that from the param list, otherwise brms will complain.
}
\details{
Importance notice that if you model lambdap1half as a sub-model, be aware that on logit scale, the non-informative prior
for intercept is logistic(0, 1) and NOT flat.
}
\seealso{
\link[brms]{prior}
}
